+++ hector/hector/doctype/transit_issue/transit_issue.js
@@ -3,7 +3,7 @@
 
 frappe.ui.form.on('Transit Issue', {
 	refresh: function(frm) {
-        if(frm.doc.workflow_state == 'Pending for Finance Team Approval' || frm.doc.workflow_state == 'Resent for Finance Team Approval')
+        if((frm.doc.workflow_state == 'Pending for Finance Team Approval' || frm.doc.workflow_state == 'Resent for Finance Team Approval') && frappe.user_roles.includes('Finance Team'))
 	    {
 	        frm.set_intro('Please Enter Credit Details');
 	    }